kafka/requirements.txt
kafka-python==2.0.2
requests
python-dotenv==1.0.0
cassandra-driver
schedule